함수 정의, 검색 설정, 페이지 이동
필수 내용을 함수로 설정
전체 검색과 상세 검색 분류
필요에 따라 검색 창을 여러 번 이용 가능
페이지 넘기는 부분은 미완성이므로 주석 처리
